requirements.txt
beautifulsoup4>=4.11.2
mysql-connector-python>=8.0.32
python-dateutil>=2.8.2
requests>=2.28.2
